set all the simulation parts to double for precision

+++ examples/05_2d_game.cpp
@@ -27,7 +27,7 @@ int main() {
   aphys::Points points = aphys::create_points();
   aphys::Lines lines = aphys::create_lines();
 
-  aphys::Matx2f v_p;
+  aphys::Matx2d v_p;
   v_p.resize(2, 2);
   v_p << 1.0f, 1.0f,  //
       1.0f, 0.5f;     //
@@ -35,37 +35,37 @@ int main() {
   v_color.resize(2, 3);
   v_color << 0.0f, 0.0f, 0.0f,  //
       1.0f, 0.2f, 0.0f;         //
-  aphys::Matx2f v_p_ref = v_p;
-  float l0 = (v_p.row(0) - v_p.row(1)).norm();
-  aphys::Matx2f v_vel;
+  aphys::Matx2d v_p_ref = v_p;
+  double l0 = (v_p.row(0) - v_p.row(1)).norm();
+  aphys::Matx2d v_vel;
   v_vel.resize(2, 2);
   v_vel << 0.0f, 0.0f,  //
       0.0f, 0.0f;       //
-  aphys::set_points_data(points, v_p, v_color);
+  aphys::set_points_data(points, v_p.cast<float>(), v_color);
   points.color = aphys::RGB(255, 0, 0);
   points.point_size = 10.0f;
-  aphys::set_lines_data(lines, v_p, v_color);
+  aphys::set_lines_data(lines, v_p.cast<float>(), v_color);
   lines.color = aphys::RGB(255, 0, 0);
   lines.width = 1.0f;
 
   aphys::Vec2f gravity(0.0f, -10.0f);
-  float mass = 1.0f;
-  float stiffness = 50.0f;
-  float damping = 0.5f;
+  double mass = 1.0;
+  double stiffness = 50.0;
+  double damping = 0.5;
 
   auto reset = [&]() {
     v_p = v_p_ref;
     v_vel.setZero();
-    aphys::set_points_data(points, v_p, v_color);
-    aphys::set_lines_data(lines, v_p, v_color);
+    aphys::set_points_data(points, v_p.cast<float>(), v_color);
+    aphys::set_lines_data(lines, v_p.cast<float>(), v_color);
   };
 
   aphys::add_gui_func(gui, [&gravity, &stiffness, &mass, &damping, &reset]() {
     ImGui::Text("Parameters:");
     ImGui::InputFloat2("g", gravity.data());
-    ImGui::InputFloat("k", &stiffness);
-    ImGui::InputFloat("m", &mass);
-    ImGui::InputFloat("beta", &damping);
+    ImGui::InputDouble("k", &stiffness);
+    ImGui::InputDouble("m", &mass);
+    ImGui::InputDouble("beta", &damping);
     if (ImGui::Button("Reset")) {
       reset();
     }
@@ -75,27 +75,28 @@ int main() {
 
   glfwSwapInterval(1);
 
-  float prev_time = glfwGetTime();
-  float start_time = prev_time;
+  double prev_time = glfwGetTime();
+  double start_time = prev_time;
 
   while (!glfwWindowShouldClose(window)) {
     aphys::set_background_RGB({244, 244, 244});
 
-    float delta_time = glfwGetTime() - prev_time;
+    double delta_time = glfwGetTime() - prev_time;
     prev_time = glfwGetTime();
-    float dt = delta_time / 10.0f;
+    double dt = delta_time / 10.0f;
 
     for (int _ = 0; _ < 10; _++) {
-      aphys::Vec2f dx = v_p.row(0) - v_p.row(1);
-      aphys::Vec2f force = gravity * mass + stiffness * (dx.norm() - l0) * dx;
-      aphys::Vec2f acc = force / mass;
+      aphys::Vec2d dx = v_p.row(0) - v_p.row(1);
+      aphys::Vec2d force =
+          gravity.cast<double>() * mass + stiffness * (dx.norm() - l0) * dx;
+      aphys::Vec2d acc = force / mass;
       v_vel.row(1).transpose() += acc * dt;
       v_vel *= exp(-damping * dt);
       v_p += v_vel * dt;
     }
 
-    aphys::set_points_data(points, v_p, v_color);
-    aphys::set_lines_data(lines, v_p, v_color);
+    aphys::set_points_data(points, v_p.cast<float>(), v_color);
+    aphys::set_lines_data(lines, v_p.cast<float>(), v_color);
 
     // aphys::Mat2fToList3f(v_p, pos);
     aphys::render_scene(scene);
