+++ include/ArmorerPhys/geom.cpp
@@ -28,16 +28,16 @@ MatxXf get_normals(const MatxXf &vertices, const MatxXi &indices) {
   return normals;
 }
 
-void create_rectangle(float l, float r, int hori_count, float b, float t,
-                      int vert_count, MatxXf &vertices, Matx3i &indices) {
+void create_rectangle(double l, double r, int hori_count, double b, double t,
+                      int vert_count, MatxXd &vertices, Matx3i &indices) {
   assert(hori_count > 0);
   assert(vert_count > 0);
   assert(l < r);
   assert(b < t);
   vertices.resize((hori_count + 1) * (vert_count + 1), 2);
   indices.resize(hori_count * vert_count * 2, 3);
-  float dx = (r - l) / hori_count;
-  float dy = (t - b) / vert_count;
+  double dx = (r - l) / hori_count;
+  double dy = (t - b) / vert_count;
   for (int i = 0; i < hori_count + 1; ++i) {
     for (int j = 0; j < vert_count + 1; ++j) {
       vertices.row(i * (vert_count + 1) + j) << l + i * dx, b + j * dy;
